Updated all migrations to check for existence before failing

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,24 +13,52 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create(
-            'service_user_t',
-            function ( Blueprint $table )
-            {
-                $table->increments( 'id' );
-                $table->string( 'first_name_text', 64 );
-                $table->string( 'last_name_text', 64 );
-                $table->string( 'email_addr_text', 320 )->unique();
-                $table->string( 'password_text', 200 );
-                $table->integer( 'owner_id' );
-                $table->integer( 'owner_type_nbr' );
-                $table->dateTime( 'last_login_date' );
-                $table->string( 'last_login_ip_text', 64 );
-                $table->string( 'remember_token', 128 );
-                $table->dateTime( 'create_date' );
-                $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
-            }
-        );
+        if ( \Schema::hasTable( 'service_user_t' ) )
+        {
+            Schema::create(
+                'service_user_t',
+                function ( Blueprint $table )
+                {
+                    $table->increments( 'id' );
+                    $table->string( 'first_name_text', 64 );
+                    $table->string( 'last_name_text', 64 );
+                    $table->string( 'email_addr_text', 320 )->unique();
+                    $table->string( 'password_text', 200 );
+                    $table->integer( 'owner_id' )->nullable();
+                    $table->integer( 'owner_type_nbr' )->nullable();
+                    $table->dateTime( 'last_login_date' )->nullable();
+                    $table->string( 'last_login_ip_text', 64 )->nullable();
+                    $table->string( 'remember_token', 128 )->nullable();
+                    $table->tinyInteger( 'active_ind' )->default( 0 );
+                    $table->dateTime( 'create_date' );
+                    $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
+                }
+            );
+        }
+
+        if ( \Schema::hasTable( 'user_t' ) )
+        {
+            Schema::create(
+                'service_user_t',
+                function ( Blueprint $table )
+                {
+                    $table->increments( 'id' );
+                    $table->string( 'first_name_text', 64 );
+                    $table->string( 'last_name_text', 64 );
+                    $table->string( 'email_addr_text', 320 )->unique();
+                    $table->string( 'password_text', 200 );
+                    $table->integer( 'owner_id' )->nullable();
+                    $table->integer( 'owner_type_nbr' )->nullable();
+                    $table->dateTime( 'last_login_date' )->nullable();
+                    $table->string( 'last_login_ip_text', 64 )->nullable();
+                    $table->string( 'remember_token', 128 )->nullable();
+                    $table->tinyInteger( 'active_ind' )->default( 0 );
+                    $table->dateTime( 'create_date' );
+                    $table->timestamp( 'lmod_date' )->default( \DB::raw( 'CURRENT_TIMESTAMP' ) );
+                }
+            );
+        }
+
     }
 
     /**
@@ -40,7 +68,15 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop( 'service_user_t' );
+        if ( \Schema::hasTable( 'service_user_t' ) )
+        {
+            Schema::drop( 'service_user_t' );
+        }
+
+        if ( \Schema::hasTable( 'user_t' ) )
+        {
+            Schema::drop( 'user_t' );
+        }
     }
 
 }
